Initial Code
Just looking at the visibility API and returning the proper logs via
the console when the visibility changes. Next steps are to integrate
with Kaltlura Plugin and change console logs to event Notifications
targeted at player. Also need to add function of detecting whether the
parent container of the video player is in the viewport